refactor and train file added
