+++ requirements.txt
@@ -1,0 +1,3 @@
+pandas~=2.2.2
+pytest~=8.2.2
+scikit-learn~=1.5.0
